+++ package.json
@@ -24,6 +24,7 @@
     "react-native-animatable": "^1.3.3",
     "react-native-dialog-input": "^1.0.8",
     "react-native-event-listeners": "^1.0.7",
+    "react-native-extended-stylesheet": "^0.12.0",
     "react-native-gesture-handler": "~2.1.0",
     "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-maps": "0.29.4",
